MatlabWrapper: improved auto cleanup of associated files (eg FDT files)
- also worked around some permission issues on Windows

+++ src/eegprep/eeglabcompat.py
@@ -46,21 +46,28 @@ class MatlabWrapper:
                 # passage data through a file
                 try:
                     temp_filename = tempfile.mktemp(dir=temp_dir, suffix='.set')
+                    result_filename = temp_filename + '.result.set'
                     pop_saveset(args[0], temp_filename)
                     # needs to use eval since returning struct arrays is not supported
                     self.engine.eval(f"EEG = pop_loadset('{temp_filename}');", nargout=0)
                     # TODO: marshalling of extra arguments should follow octave conventions
                     eval_str = f"EEG = {name}(EEG{',' if args[1:] else ''}{','.join([self.marshal(a) for a in args[1:]])});"
-                    print(eval_str)
+                    print(f"Running in MATLAB: {eval_str}")
                     self.engine.eval(eval_str, nargout=0)
-                    self.engine.eval(f"pop_saveset(EEG, '{temp_filename}');", nargout=0)
-                    return pop_loadset(temp_filename)
+                    self.engine.eval(f"pop_saveset(EEG, '{result_filename}');", nargout=0)
+                    return pop_loadset(result_filename)
                 finally:
                     # delete temporary file
                     try:
                         # noinspection PyUnboundLocalVariable
-                        os.remove(temp_filename)
+                        if os.path.exists(temp_filename):
+                            os.remove(temp_filename)
                         if os.path.exists(fdt_file := temp_filename.replace('.set', '.fdt')):
+                            os.remove(fdt_file)
+                        # noinspection PyUnboundLocalVariable
+                        if os.path.exists(result_filename):
+                            os.remove(result_filename)
+                        if os.path.exists(fdt_file := result_filename.replace('result.set', 'result.fdt')):
                             os.remove(fdt_file)
                     except OSError as e:
                         logger.warning(f"Error deleting temporary file {temp_filename}: {e}")
